fix(music): the album lookup asked AcoustID for a shape it could not read
Adding music from a folder returned "No results" for anything without embedded
MusicBrainz tags. The filelist asks AcoustID for meta=recordings+releases+
releasegroups, and requesting releasegroups alongside releases makes AcoustID
move the releases down into recordings[].releasegroups[].releases[]. All three
fingerprint consumers read recordings[].releases[], so every one of them found
an empty array, looked nothing up, and reported success. The request logged as
ok in 8.4s with zero entries and no MusicBrainz call to show for it.

Drops releasegroups from the lookup. The same Blondie fingerprints that yielded
0 release ids now yield 223-283, resolving to the 20-track compilations the
folder actually holds.

That fan-out then had to be bounded: FromFingerprint fetched every named release
from a rate-limited MusicBrainz and then ffprobed the whole folder to score each
one, which for three tracks alone meant 324 distinct releases. Candidates are
now collected from every track first and ranked before anything is fetched, on
the two facts that decide it: a release whose track count differs from the file
count cannot be this folder, and the real album is the one most of the tracks
name. It also stops abandoning a file's remaining candidates when one release
fails to fetch.

Recording duration was typed int against a float wire value ("duration":
205.291), so the bind failed, the global Newtonsoft error handler swallowed it,
and MusicLogic's duration match compared every track against 0.

+++ src/NoMercy.Data/Services/MusicLogic.cs
@@ -333,7 +333,7 @@ public partial class MusicLogic : IAsyncDisposable
 
             // var ffProbeData = FFProbe.AnalyseAsync(file.Path).Result;
             double fileDuration = ffProbeData.Format.Duration.TotalSeconds;
-            int recordDuration = recording.Duration;
+            double recordDuration = recording.Duration;
             bool matchesDuration =
                 fileDuration > 0
                 && recordDuration > 0
